Decrease runtime memory overhead by getting rid of AllocationInfo.
We now output every allocation's ptr and only output it for a free.
This not only reduces the runtime CPU overhead, but also drastically
reduces the memory overhead as the relatively large AllocationInfo
map is not required anymore.

The size of the output data files increases, but the cost seems to
be bearable and worth the reduces memory overhead imo.