Merge allocation data by instruction pointer.
This is safe to do except for the peak memory consumption where
we cannot ignore the temporal dependency.

Otherwise this greatly improves the output as multiple calls to the
same function from different places get merged and thus show up
as one fat data point, instead of multiple thinner ones.