Refactor hook initialization in heaptrack_preload.
The previous code works with clang, but not with GCC. This new
code works with GCC, and, hopefully, also with clang. It does not
depend on static initialization order anymore, as the string
identifiers of the hook functions are initialized via constexpr now.

This fixes the heaptrack_preload usage for me on this machine.