Use a custom thread + sleep_for instead of a C timer.
This makes re-attaching work without deadlocks in timer_delete
for me. And we can clean the code up even more now:

Much less code, yet still more accurate. The overhead is negleglible
in my tests. And this makes it trivial for the future to let users
configure the interval themselves.