Also sum up peaks.
The models have no temporal notation thus if two peaks happen
at a different time, you cannot say that this is the real
peak memory consumption of a merged allocation.

But taking the absolute peak of peaks is very confusing as
then the leaked number can be less than the peak.