Introduce cost threshold for flamegraph to minimize memory overhead.
For larger data files, the flamegraph can easily consume hundreds
of megabytes of ram and take multiple seconds to be build up. This
is mostly due to the dozens of small leaf items, all of which only
have tiny cost and will not be shown most of the time.

The new default cut-off of 0.1% drastically reduces the memory
overhead and the flamegraph will be build nearly instantly even for
larger data sets.

In the future, one could even refine the data on-demand.