Move libunwind CMake logic into FindLibunwind.cmake
This fixes a build regression when using a clean build folder.

The find script is copied from zbackup, cf.:

https://raw.githubusercontent.com/zbackup/zbackup/master/cmake/FindLibUnwind.cmake

It was slightly adapted to find the symbols required for heaptrack
and not depend on symbols we don't actually need here.

While at it, extract the configuration defines out into its own
libunwind_config.h.