Elide text of flamegraph items
This prevents the flickering when hovering an item with a long text.
That lead the label to wordwrap, increasing the height of the label,
decreasing the height of the view, moving the scene, thus the mouse
points to a different view and we would repeat again.

Now we simply elide the text on the right of the label. This works,
but it's cutting of the text too soon, since it cannot cope with
the HTML and thinks all of these chars will get painted...