Fix detection of unw_set_cache_size and adapt to upstream changes
unw_set_cache_size is actually a define so we can detect it directly
in C++ without the need for CMake to detect it for us. Using the
define also makes sure we can detect it once for all platforms, the
symbol is mangled by libunwind to support multiple architectures.