Optimize aggregation of caller/callee data
Previously, we did that on the unmerged data which means we had
to traverse way more stacks. Now, we instead use the merged data
as input and thereby can get rid of ~50% of the hash allocations
alone used for the recursion guard. This is the #1 hotspot when
it comes to allocation numbers so far, so that boils down to easily
millions of allocations.