Prefer $lib.debug files over $lib files, if available
This is the first step towards supporting split-debug files. This
patch alone will at least improve the situation when you are using
heaptrack on a binary that has split debug information in a file
with the ".debug" suffix. In such cases, we used to be unable to
resolve the file/line information. Now, we properly resolve those
too. Note that Qt 5 QMake uses this split file configuration by
default since some time.