Make run_linkage_test.sh pass
When we read a function pointer before attaching heaptrack,
that value won't get rewritten. But when we read the pointer
address later on, the correct value can be found and the
test passes as expected.

I don't think there's anything we can do about the first
situation - we don't want to rewrite the whole address space...