Simplify enabling of sanitizers for non-injected heaptrack code
Sanitizing the injected heaptrack code is not possible to my
knowledge, since asan at least will interfer with us overriding
malloc. I'll have to figure out how this can be done separately.

For now, this patch just adds a trivial way to configure heaptrack
to be build with sanitizers via a cache variable. This is then
applied to all of the non-injected code, namely heaptrack_interpret,
heaptrack_print and heaptrack_gui.