Optimize AccumulatedTraceData::findAllocation
Instead of sorting the vector of Allocation objects which are
48byte large, introduce a separate sorted vector of pairs of
TraceIndex and and AllocationIndex, both just 4 byte large.
Lookup and insertion in the middle of this much smaller container
is considerably faster, improving the heaptrack_print analysis
time by ~10% in one of my larger test files (from ~3minutes down
to 2min40s).

CCBUG: 386256