Do not use shell to launch new programs in debug mode
When we want to debug heaptrack we use GDB to launch the new
process. By default, that actually first starts a shell and then
eventually the debuggee. But both processes would get heaptrack
injected and will then write to the same heaptrack_interpret
process, thereby corrupting the data file.

By using `set startup-with-shell off` in GDB, we can eliminate the
shell and prevent this issue from occurring.

BUG: 392085