Add basic unit test coverage for runtime attaching
This test will use dlopen or the equivalent internal libc API
to load heaptrack, trigger an allocation, and then stop and
unload the library.

Note that we don't actually test whether the allocation was
tracked yet, since that actually doesn't work in this scenario
here yet. I'm investigating what is happening...