Add scripts to generate heaptrack AppImages in Docker
The build_appimage.sh script generates an AppImage, it could
eventually be run directly on the KDE CI.

The build_appimage_in_docker.sh script takes the Dockerfile and
builds a container that's ready to run the above build_appimage.sh
script. The generated heaptrack AppImage is then copied to the
current folder.