Call libc / libstdc++ freeres function in libheaptrack's atexit
On one hand, this prevents an infinite recursion when the call to
atexit allocates, which happens e.g. with uClibc, since the
call from heaptrack_preload did not install the recursion guard yet.

On the other hand, this code move ensures that the freeres functions
get called properly before heaptrack_stop is called.

BUG: 395109