Allow for writing arbitrary long strings in LineWriter
The write function that takes a formatted string only supports
strings up to BUFFER_CAPACITY size. When we write symbol strings in
heaptrack_interpret, these can be arbitrarily large thanks to C++
template magic.

For this case, introduce a new overload that prints a std::string
directly, bypassing the buffer.