Add context menu entry to jump to caller/callee view from flame graph
This action greatly simplifies the analysis routine when using the
flame graph. Previously, one had to manually search for a given
entry in the caller/callee view to find the source locations. Now,
we can just use the context menu to jump to the caller/callee view
and open the editor for a given source location from its source map
view.

To do so, we have to aggregate flame graph items by symbol, instead
of just by the function name. This is useful anyhow, since now we
can see what DSO a given function correlates to.