Checks if leaked memory resident in RAM.
Previously all virtual memory that is not released is considered as a leak.
This is a problem in most of the cases but not all.
Sometimes we would like to see only physical memory (RAM) which leaks. E.g.:
int* a = new int[1024*1024*1024];
int* b = new int[1024];
for(size_t i = 0; i < 1024; ++i)
    b[1024] = 0xdeadbeef;
"b" is more worrying than "a" since "a" does not leak RAM (at least on Linux by defaul).
"b" - requires one or more pages of RAM. When "a" does not use RAM at all (https://marek.vavrusa.com/memory/).
If you care more about RAM than virtual memory you should take a look.

The change introduce category "In RAM" which shows memory that leaks and is present in RAM.