Provide progress information during the "merging allocations" phase
Summary:
It can take a long time, and I needed the progress information to see
if I had enough RAM to let it finish or if it was never going to work.

Interestingly, this shows that the same loop becomes slower and slower.

Test Plan: heaptrack kmail, read 4-5 emails, quit

Reviewers: mwolff

Reviewed By: mwolff

Differential Revision: https://phabricator.kde.org/D24088