Implement time range filtering in chart widgets
Seems to work, but the UI needs to be adapted to show the currently
zoomed time range somewhere in addition to a button to undo the filter
easily.

Also for large detailed graphs the painting is quite expensive,
we should cache that for the chart and only redo it when needed.