Use a cached pixmap while interacting with the chart rubber band
This has a huge impact on the perceived performance of selecting a
time range. Previously, it was super slow as KDiagram is just very
slow at reparsing large charts. Now, it's butter smooth but slightly
too bright compared to the normal way of drawing the chart. But we
only do that while resizing the rubber band, so I think that's fine.