Fix assertion in AccumulatedTraceData::mapToAllocationIndex
When we encounter trace indices in the order A, B, A, where
the index A is the current maximum but B is less than that,
we incorrectly mapped the second occurrence of A to B.

Fix this by explicitly remembering the allocation index of the
maximum index. This allows us to keep the fast-path code and
at the same time prevents the improperly attribution.

BUG: 427970