Unbreak tst_inject in release mode
For some very strange reason, the `throw LockCheckFailed` directly
lead to a call to std::terminate there. I have not yet understood
why this is the case, but I suspect it's due to heaptrack intercepting
some memory allocation that happens during unwinding? Anyhow, when
we throw+catch once earlier, then the issue seems to disappear. So
let's do that to prevent this very odd breakage...