Further optimize recursion guard handling
Replace lookup+insert with just an insert and check if the size
changed to track whether we actually overwrote an item or not.
Sadly, QSet doesn't have the nice API of std::unordered_set here,
but otherwise it's faster.

Also use SymbolId in one more place.

Sadly, the performance impact isn't really noticable.