Add basic support for leak suppressions to heaptrack_print
When the new `--suppressions <file>` parameter is supplied,
the given file will be parsed to create a list of suppression
rules. The format is expected to be LSAN compatible, but for
now only simple substring matching is supported without anchoring.

Then during the finalize step, we efficiently check which allocation
events match the suppressions and set their leak cost to 0. As such,
no other changes are required and it will be easy to apply the same
approach to heaptrack_gui next.

The suppression strings are matched against library/executable names,
function names as well as source file names. Line numbers are not
supported in the leak suppression file.

The final summary printed will show whether any leaks got suppressed.

Work on this patch was sponsored by NetworkRADIUS.com - thanks!