Reuse the TemplateMatch suppression function from the sanitizers
Doing so adds support for anchoring and globbing in the suppressions.

This makes sure that our behavior is identical to whatever the
sanitizers are doing. I have to admit though, that code does look
a bit fishy, but it certainly works.

Work on this patch was sponsored by NetworkRADIUS.com - thanks!