Optional allow using heaptrack_inject via LD_PRELOAD
When you now invoke `heaptrack --use-inject app`, we will load
the `heaptrack_inject.so` instead of `heaptrack_preload.so` and
rely on the symbol interception logic therein. This can yield better
results in at least one test case, were the other approach that relies
on LD_PRELOAD and the dynamic linker to resolve `free` and friends
fails to intercept _some_ calls to `free`. The other approach that
we employ in `heaptrack_inject.so` works though.

I'm not yet confident in saying which approach is better, i.e. whether
we should always use `heaptrack_inject.so`. There may be some cases
that work better with the other approach, but that will only show
with more testing.

For now, adding this feature behind an optional command line switch
is the best I can offer, until I manage to fully grasp why the
dynamic linker and LD_PRELOAD sometimes fails.

Work on this patch was sponsored by NetworkRADIUS.com - thanks!