Fix diffing algorithm
The previous implementation was seriously broken and I have to admit
I'm a bit ashamed at not realizing this before...

It merged items from within the erase_if callback, and tried to
remember the position of the items at that stage too. But once the
erase happened, the positions where totally off, meaning the results
produced pretty random broken wrong results. Whops.

Now, we first sort, then cleanup, and generally restructure the code
to not rely on positions within the allocations list anymore.

The code is actually much nicer now, no temporary other containers
are required, we can just sort the results directly and then handle
it from there on.