Warn when detecting dlopen(RTLD_DEEPBIND) and disable it
When we use libheaptrack_preload, dlopen with RTLD_DEEPBIND would
lead to us not being able to intercept calls to memory (de-)allocation
functions. Detect this, output a warning, and then disable that
flag to allow us to intercept all functions as desired.

Applications that rely on this mechanism need to pass `--use-inject`
as that seems to work correctly even with RTLD_DEEPBIND.

Work on this patch was sponsored by NetworkRADIUS.com - thanks!