Fix backtrace symbolication for optimized code
For some reason, we actually get an IP address that points to the
instruction _after_ the one we need to use for symbolication. I'm
quite frankly totally flabbergasted by this breakage, as it seems
to be around for years, yet no-one ever complained? And testing this
on some of my code, I believe it used to work in the past without
this workaround... If anyone knows more about this, please chime
in over at: https://github.com/libunwind/libunwind/issues/287

For now, manually subtracting one from the addresses we get from
unwinding is enough to workaround this and get sane backtraces once
more.

BUG: 439897