Fix latent bug in ChartData creation
We created too many labeled values, leading to an overflow
of the ChartRows::cost array. Previously, this assert didn't
hit until filtering, because the model checked the size of the
wrong container.

This patch fixes this all: We now ensure we stay within
ChartRows::MAX_NUM_COST but also increase it by one to keep the
same size as before. Furthermore, we now correctly assert the size
of the right container, and additionally check this already during
parsing and not only later when the model is set.

BUG: 441132