Add copy of dwarf.h with support for DWARF5
The newer libbacktrace version we are shipping is using newer
defines from that header which are not yet available on the dwarf.h
shipped with Centos 7. Thankfully, this header is nicely portable
and self-contained, and licensing (LGPL2.1) should allow shipping
it with our source code too.

Obviously, this now means we'll have to update both, libbacktrace
and libdwarf. But I plan to port heaptrack to use elfutils soon
anyways. At that point, we will be able to remove both again.

BUG: 441125