Let releaseme know to skip po install
Since KF5I18n is optional AND dependent on HEAPTRACK_BUILD_GUI, we
have to call ki18n_install(po) conditionally on our own and in
addition tell releaseme not to touch CMakeLists.txt to do the same.

Otherwise releaseme adds the following code unconditionally at the
end of the file:

    find_package(KF5I18n CONFIG REQUIRED)
    ki18n_install(po)

ki18n_install checks for existence of po subdir and returns silently
if none found, so it is safe for git sources too.

Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>