Don't use exceptions when timer thread fails to lock
This is a bit unweildy without C++17 and std::optional, but it gets
the job done. Apparently throwing exceptions doesn't work from within
heaptrack when it's injected into a flutter binary, for unclear
reasons. Maybe due to [1] but that's supposedly fixed in newer flutter
versions already. But as soon as we throw, we would terminate in the
context of a terminate.

[1]: https://github.com/flutter/flutter/issues/66575