Use boost::hash_combine to prevent collisions in tsl pointermap
Apparently the hash combine I copied from stackoverflow all these
years ago is simply broken - it can trigger a worst-case scenario
in the tsl::set leading to excessive memory consumption.

Using boost::hash_combine the example given at [1] we get back to
a load factor of ~0.25 at the end, instead of going all the way to
0.00... and gigabytes of memory consumed for the buckets.

We use boost already elsewhere so it's fine to use it here too.

[1]: https://github.com/Tessil/robin-map/issues/71