Handle Damaged Timestamps in Files
If a heaptrack output file is copied while the application is still running, the timestamp may become corrupted. However, accessing the file before the application ends can still be useful for testing. To address this, the case of a damaged timestamp should be handled. Currently, if the last timestamp is corrupted, a file that should span multiple hours is displayed as only a few seconds long. By ignoring timestamps smaller than the last valid one, this issue is resolved, and the full file is correctly presented in the GUI or in print.