Check to resolve files using their full path against --extra-paths
The paths passed with --extra-paths are treated as the full path to a
directory containing debug symbols; but for very large projects which
deploy executables and shared objects together, this can be cumbersome
to hunt down and list all those paths.

If such a project is deployed from a local staging directory (e.g. by
running DESTDIR=./staging make install), this change will cause
heaptrack to find debug symbols if that staging directory is given as an
extra path.

Example:
Given the following directory tree

staging
  opt
    myproject
      bin
        myexecutable
        .debug
      lib
        somelib.so
        otherlib.so
        .debug
      qml
        qmlplugin.so
        .debug

Assuming that the files are deployed with something like
    rsync staging/opt remotehost:/

It would be necessary to interpret a raw heaptrack report using 3 extra
paths, one for each of the .debug folders.

With this change, it works to use just one extra path pointing to the
staging directory, because the relative paths to the files on the device
are same as they are relative to the staging directory.